Unsafe change: Yanked out direct cassandra interaction in CassandraAccessLayer.
Note that this isn't a refactoring, since I don't have tests. So who knows
if this works. I've been careful :)

Overall, I've moved all interactions with the concrete class Cluster
into a small wrapper with an interface. This reduces the surface of
cluster we expose, and allows us to mock the cluster and the
session, so we can test the cassandra access layer here.